+++ crates/mesh-llm-host-runtime/src/network/openai/moa_gateway/mod.rs
@@ -106,6 +106,7 @@ pub async fn try_handle_moa(
 }
 
 pub(in crate::network::openai) mod context_selection;
+mod pool;
 mod progress;
 mod streaming;
 mod workers;
